Split some code out into btree_error
