+++ src/pdata/array_walker.rs
@@ -4,7 +4,8 @@ use std::sync::{Arc, Mutex};
 use crate::checksum;
 use crate::io_engine::*;
 use crate::pdata::array::{self, *};
-use crate::pdata::btree::{self, *};
+use crate::pdata::btree::*;
+use crate::pdata::btree_error::{self, *};
 use crate::pdata::btree_walker::*;
 use crate::pdata::space_map::*;
 use crate::pdata::unpack::*;
@@ -70,7 +71,7 @@ impl<'a, V: Unpack> NodeVisitor<u64> for BlockValueVisitor<'a, V> {
         _h: &NodeHeader,
         keys: &[u64],
         values: &[u64],
-    ) -> btree::Result<()> {
+    ) -> btree_error::Result<()> {
         if keys.is_empty() {
             return Ok(());
         }
@@ -79,17 +80,17 @@ impl<'a, V: Unpack> NodeVisitor<u64> for BlockValueVisitor<'a, V> {
         // The ordering of keys had been verified in unpack_node(),
         // so comparing the keys against its context is sufficient.
         if *keys.first().unwrap() + keys.len() as u64 != *keys.last().unwrap() + 1 {
-            return Err(btree::value_err("gaps in array indicies".to_string()));
+            return Err(btree_error::value_err("gaps in array indicies".to_string()));
         }
         if let Some(end) = kr.end {
             if *keys.last().unwrap() + 1 != end {
-                return Err(btree::value_err(
+                return Err(btree_error::value_err(
                     "non-contiguous array indicies".to_string(),
                 ));
             }
         }
 
-        let mut errs = Vec::<btree::BTreeError>::new();
+        let mut errs = Vec::<btree_error::BTreeError>::new();
         match self.engine.read_many(values) {
             Err(_) => {
                 // IO completely failed on all the child blocks
@@ -124,11 +125,11 @@ impl<'a, V: Unpack> NodeVisitor<u64> for BlockValueVisitor<'a, V> {
         }
     }
 
-    fn visit_again(&self, _path: &[u64], _b: u64) -> btree::Result<()> {
+    fn visit_again(&self, _path: &[u64], _b: u64) -> btree_error::Result<()> {
         Ok(())
     }
 
-    fn end_walk(&self) -> btree::Result<()> {
+    fn end_walk(&self) -> btree_error::Result<()> {
         Ok(())
     }
 }
@@ -163,7 +164,7 @@ impl ArrayWalker {
         })
     }
 
-    pub fn walk<V>(&self, visitor: &dyn ArrayVisitor<V>, root: u64) -> btree::Result<()>
+    pub fn walk<V>(&self, visitor: &dyn ArrayVisitor<V>, root: u64) -> btree_error::Result<()>
     where
         V: Unpack,
     {
@@ -197,16 +198,16 @@ impl NodeVisitor<u64> for BlockPathCollector {
         _h: &NodeHeader,
         keys: &[u64],
         values: &[u64],
-    ) -> btree::Result<()> {
+    ) -> btree_error::Result<()> {
         // Verify key's continuity.
         // The ordering of keys had been verified in unpack_node(),
         // so comparing the keys against the key range is sufficient.
         if *keys.first().unwrap() + keys.len() as u64 != *keys.last().unwrap() + 1 {
-            return Err(btree::value_err("gaps in array indicies".to_string()));
+            return Err(btree_error::value_err("gaps in array indicies".to_string()));
         }
         if let Some(end) = kr.end {
             if *keys.last().unwrap() + 1 != end {
-                return Err(btree::value_err(
+                return Err(btree_error::value_err(
                     "non-contiguous array indicies".to_string(),
                 ));
             }
@@ -220,11 +221,11 @@ impl NodeVisitor<u64> for BlockPathCollector {
         Ok(())
     }
 
-    fn visit_again(&self, _path: &[u64], _b: u64) -> btree::Result<()> {
+    fn visit_again(&self, _path: &[u64], _b: u64) -> btree_error::Result<()> {
         Ok(())
     }
 
-    fn end_walk(&self) -> btree::Result<()> {
+    fn end_walk(&self) -> btree_error::Result<()> {
         Ok(())
     }
 }
@@ -233,7 +234,7 @@ pub fn collect_array_blocks_with_path(
     engine: Arc<dyn IoEngine + Send + Sync>,
     ignore_non_fatal: bool,
     root: u64,
-) -> btree::Result<BTreeMap<u64, (Vec<u64>, u64)>> {
+) -> btree_error::Result<BTreeMap<u64, (Vec<u64>, u64)>> {
     let walker = BTreeWalker::new(engine, ignore_non_fatal);
     let visitor = BlockPathCollector::new();
     walker.walk(&mut vec![0], &visitor, root)?;
