+++ src/pdata/bitset.rs
@@ -5,7 +5,7 @@ use crate::io_engine::IoEngine;
 use crate::math::div_up;
 use crate::pdata::array::{self, ArrayBlock};
 use crate::pdata::array_walker::{ArrayVisitor, ArrayWalker};
-use crate::pdata::btree;
+use crate::pdata::btree_error;
 use crate::pdata::space_map::*;
 
 //------------------------------------------
@@ -140,7 +140,7 @@ pub fn read_bitset_checked(
     root: u64,
     nr_bits: usize,
     ignore_none_fatal: bool,
-) -> (CheckedBitSet, Option<btree::BTreeError>) {
+) -> (CheckedBitSet, Option<btree_error::BTreeError>) {
     let w = ArrayWalker::new(engine, ignore_none_fatal);
     let v = BitsetVisitor::new(nr_bits);
     let err = match w.walk(&v, root) {
@@ -157,7 +157,7 @@ pub fn read_bitset_checked_with_sm(
     nr_bits: usize,
     sm: Arc<Mutex<dyn SpaceMap + Send + Sync>>,
     ignore_none_fatal: bool,
-) -> array::Result<(CheckedBitSet, Option<btree::BTreeError>)> {
+) -> array::Result<(CheckedBitSet, Option<btree_error::BTreeError>)> {
     let w = ArrayWalker::new_with_sm(engine, sm, ignore_none_fatal)?;
     let v = BitsetVisitor::new(nr_bits);
     let err = match w.walk(&v, root) {
@@ -172,7 +172,7 @@ pub fn read_bitset(
     root: u64,
     nr_bits: usize,
     ignore_none_fatal: bool,
-) -> btree::Result<FixedBitSet> {
+) -> btree_error::Result<FixedBitSet> {
     let w = ArrayWalker::new(engine, ignore_none_fatal);
     let v = BitsetCollector::new(nr_bits);
     w.walk(&v, root)?;
