+++ src/pdata/mod.rs
@@ -4,6 +4,7 @@ pub mod array_walker;
 pub mod bitset;
 pub mod btree;
 pub mod btree_builder;
+pub mod btree_error;
 pub mod btree_leaf_walker;
 pub mod btree_merge;
 pub mod btree_walker;
